// AutoVideoMetaLocalize/client/generated-sources/openapi/models/ChannelListResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AutoVideoMetaLocalize
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Channel,
    ChannelFromJSON,
    ChannelFromJSONTyped,
    ChannelToJSON,
    PageInfo,
    PageInfoFromJSON,
    PageInfoFromJSONTyped,
    PageInfoToJSON,
    TokenPagination,
    TokenPaginationFromJSON,
    TokenPaginationFromJSONTyped,
    TokenPaginationToJSON,
} from './';

/**
 * 
 * @export
 * @interface ChannelListResponse
 */
export interface ChannelListResponse {
    /**
     * 
     * @type {string}
     * @memberof ChannelListResponse
     */
    eTag?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ChannelListResponse
     */
    eventId?: string | null;
    /**
     * 
     * @type {Array<Channel>}
     * @memberof ChannelListResponse
     */
    items?: Array<Channel> | null;
    /**
     * 
     * @type {string}
     * @memberof ChannelListResponse
     */
    kind?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ChannelListResponse
     */
    nextPageToken?: string | null;
    /**
     * 
     * @type {PageInfo}
     * @memberof ChannelListResponse
     */
    pageInfo?: PageInfo;
    /**
     * 
     * @type {string}
     * @memberof ChannelListResponse
     */
    prevPageToken?: string | null;
    /**
     * 
     * @type {TokenPagination}
     * @memberof ChannelListResponse
     */
    tokenPagination?: TokenPagination;
    /**
     * 
     * @type {string}
     * @memberof ChannelListResponse
     */
    visitorId?: string | null;
}

export function ChannelListResponseFromJSON(json: any): ChannelListResponse {
    return ChannelListResponseFromJSONTyped(json, false);
}

export function ChannelListResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChannelListResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'eTag': !exists(json, 'eTag') ? undefined : json['eTag'],
        'eventId': !exists(json, 'eventId') ? undefined : json['eventId'],
        'items': !exists(json, 'items') ? undefined : (json['items'] === null ? null : (json['items'] as Array<any>).map(ChannelFromJSON)),
        'kind': !exists(json, 'kind') ? undefined : json['kind'],
        'nextPageToken': !exists(json, 'nextPageToken') ? undefined : json['nextPageToken'],
        'pageInfo': !exists(json, 'pageInfo') ? undefined : PageInfoFromJSON(json['pageInfo']),
        'prevPageToken': !exists(json, 'prevPageToken') ? undefined : json['prevPageToken'],
        'tokenPagination': !exists(json, 'tokenPagination') ? undefined : TokenPaginationFromJSON(json['tokenPagination']),
        'visitorId': !exists(json, 'visitorId') ? undefined : json['visitorId'],
    };
}

export function ChannelListResponseToJSON(value?: ChannelListResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'eTag': value.eTag,
        'eventId': value.eventId,
        'items': value.items === undefined ? undefined : (value.items === null ? null : (value.items as Array<any>).map(ChannelToJSON)),
        'kind': value.kind,
        'nextPageToken': value.nextPageToken,
        'pageInfo': PageInfoToJSON(value.pageInfo),
        'prevPageToken': value.prevPageToken,
        'tokenPagination': TokenPaginationToJSON(value.tokenPagination),
        'visitorId': value.visitorId,
    };
}