// AutoVideoMetaLocalize/client/generated-sources/openapi/models/PlaylistItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AutoVideoMetaLocalize
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    PlaylistItemContentDetails,
    PlaylistItemContentDetailsFromJSON,
    PlaylistItemContentDetailsFromJSONTyped,
    PlaylistItemContentDetailsToJSON,
    PlaylistItemSnippet,
    PlaylistItemSnippetFromJSON,
    PlaylistItemSnippetFromJSONTyped,
    PlaylistItemSnippetToJSON,
    PlaylistItemStatus,
    PlaylistItemStatusFromJSON,
    PlaylistItemStatusFromJSONTyped,
    PlaylistItemStatusToJSON,
} from './';

/**
 * 
 * @export
 * @interface PlaylistItem
 */
export interface PlaylistItem {
    /**
     * 
     * @type {PlaylistItemContentDetails}
     * @memberof PlaylistItem
     */
    contentDetails?: PlaylistItemContentDetails;
    /**
     * 
     * @type {string}
     * @memberof PlaylistItem
     */
    eTag?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlaylistItem
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlaylistItem
     */
    kind?: string | null;
    /**
     * 
     * @type {PlaylistItemSnippet}
     * @memberof PlaylistItem
     */
    snippet?: PlaylistItemSnippet;
    /**
     * 
     * @type {PlaylistItemStatus}
     * @memberof PlaylistItem
     */
    status?: PlaylistItemStatus;
}

export function PlaylistItemFromJSON(json: any): PlaylistItem {
    return PlaylistItemFromJSONTyped(json, false);
}

export function PlaylistItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlaylistItem {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'contentDetails': !exists(json, 'contentDetails') ? undefined : PlaylistItemContentDetailsFromJSON(json['contentDetails']),
        'eTag': !exists(json, 'eTag') ? undefined : json['eTag'],
        'id': !exists(json, 'id') ? undefined : json['id'],
        'kind': !exists(json, 'kind') ? undefined : json['kind'],
        'snippet': !exists(json, 'snippet') ? undefined : PlaylistItemSnippetFromJSON(json['snippet']),
        'status': !exists(json, 'status') ? undefined : PlaylistItemStatusFromJSON(json['status']),
    };
}

export function PlaylistItemToJSON(value?: PlaylistItem | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'contentDetails': PlaylistItemContentDetailsToJSON(value.contentDetails),
        'eTag': value.eTag,
        'id': value.id,
        'kind': value.kind,
        'snippet': PlaylistItemSnippetToJSON(value.snippet),
        'status': PlaylistItemStatusToJSON(value.status),
    };
}